Implement Downloads page
This change includes an update to the database, adding a downloaded
column in Enclosures.

+++ src/enclosure.cpp
@@ -12,6 +12,7 @@
 #include <QSqlQuery>
 
 #include "database.h"
+#include "datamanager.h"
 #include "enclosuredownloadjob.h"
 #include "entry.h"
 #include "fetcher.h"
@@ -35,18 +36,46 @@ Enclosure::Enclosure(Entry *entry)
     m_type = query.value(QStringLiteral("type")).toString();
     m_url = query.value(QStringLiteral("url")).toString();
     m_playposition = query.value(QStringLiteral("playposition")).toLongLong();
+    m_status = query.value(QStringLiteral("downloaded")).toBool() ? Downloaded : Downloadable;
     m_playposition_dbsave = m_playposition;
 
+    // In principle the database contains this status, we check anyway in case
+    // something changed on disk
     QFile file(path());
     if (file.exists()) {
         if(file.size() == m_size) {
-            m_status = Downloaded;
+            if (m_status == Downloadable) {
+                // file is on disk, but was not expected, write to database
+                // this should never happen
+                m_status = Downloaded;
+                query.prepare(QStringLiteral("UPDATE Enclosures SET downloaded=:downloaded WHERE id=:id;"));
+                query.bindValue(QStringLiteral(":id"), entry->id());
+                query.bindValue(QStringLiteral(":downloaded"), true);
+                Database::instance().execute(query);
+            }
         } else {
-            file.remove();
-            m_status = Downloadable;
+            if (m_status == Downloaded) {
+                // file was downloaded, but there is a size mismatch
+                // delete file and update status in database
+                file.remove();
+                m_status = Downloadable;
+                query.prepare(QStringLiteral("UPDATE Enclosures SET downloaded=:downloaded WHERE id=:id;"));
+                query.bindValue(QStringLiteral(":id"), entry->id());
+                query.bindValue(QStringLiteral(":downloaded"), false);
+                Database::instance().execute(query);
+            }
         }
     } else {
-        m_status = Downloadable;
+        if (m_status == Downloaded) {
+            // file was supposed to be on disk, but isn't there
+            // update status and write to the database
+            file.remove();
+            m_status = Downloadable;
+            query.prepare(QStringLiteral("UPDATE Enclosures SET downloaded=:downloaded WHERE id=:id;"));
+            query.bindValue(QStringLiteral(":id"), entry->id());
+            query.bindValue(QStringLiteral(":downloaded"), false);
+            Database::instance().execute(query);
+        }
     }
 }
 
@@ -63,9 +92,7 @@ void Enclosure::download()
 
     connect(downloadJob, &KJob::result, this, [this, downloadJob]() {
         if(downloadJob->error() == 0) {
-            m_status = Downloaded;
             processDownloadedFile();
-
         } else {
             m_status = Downloadable;
             if(downloadJob->error() != QNetworkReply::OperationCanceledError) {
@@ -81,6 +108,7 @@ void Enclosure::download()
         downloadJob->doKill();
         m_status = Downloadable;
         Q_EMIT statusChanged();
+        Q_EMIT DataManager::instance().downloadCountChanged(m_entry->feed()->url());
         disconnect(this, &Enclosure::cancelDownload, this, nullptr);
     });
 
@@ -95,7 +123,12 @@ void Enclosure::download()
 
 void Enclosure::processDownloadedFile() {
     // This will be run if the enclosure has been downloaded successfully
-
+    m_status = Downloaded;
+    QSqlQuery query;
+    query.prepare(QStringLiteral("UPDATE Enclosures SET downloaded=:downloaded WHERE id=:id;"));
+    query.bindValue(QStringLiteral(":id"), m_entry->id());
+    query.bindValue(QStringLiteral(":downloaded"), true);
+    Database::instance().execute(query);
     // Unset "new" status of item
     if (m_entry->getNew()) m_entry->setNew(false);
 
@@ -113,6 +146,8 @@ void Enclosure::processDownloadedFile() {
         query.bindValue(QStringLiteral(":size"), m_size);
         Database::instance().execute(query);
     }
+    Q_EMIT DataManager::instance().downloadCountChanged(m_entry->feed()->url());
+
 }
 
 void Enclosure::deleteFile()
@@ -123,8 +158,13 @@ void Enclosure::deleteFile()
         QFile(path()).remove();
     // If file disappeared unexpectedly, then still change status to downloadable
     m_status = Downloadable;
+    QSqlQuery query;
+    query.prepare(QStringLiteral("UPDATE Enclosures SET downloaded=:downloaded WHERE id=:id;"));
+    query.bindValue(QStringLiteral(":id"), m_entry->id());
+    query.bindValue(QStringLiteral(":downloaded"), false);
+    Database::instance().execute(query);
     Q_EMIT statusChanged();
-
+    Q_EMIT DataManager::instance().downloadCountChanged(m_entry->feed()->url());
 }
 
 QString Enclosure::path() const
@@ -136,7 +176,6 @@ Enclosure::Status Enclosure::status() const
 {
     return m_status;
 }
-
 qint64 Enclosure::playPosition() const{
     return m_playposition;
 }
